Rename package and folders
