+++ src/main/java/com/google/impactdashboard/data/project/ProjectIdentification.java
@@ -1,4 +1,4 @@
-package com.google.sps.data;
+package com.google.impactdashboard.data.project;
 
 /** Represents the identifying information of a single project. */
 public class ProjectIdentification {
